api/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
psycopg2-binary==2.9.3
pymemcache==3.5.2 